Move multisig test to `keriox_tests`
