+++ keriox_tests/Cargo.toml
@@ -12,3 +12,6 @@ tempfile = "3.3.0"
 url = "2.2.2"
 watcher = { path = "../components/watcher" }
 witness = { path = "../components/witness" }
+controller= { path = "../components/controller" }
+async-std = { version = "1.12.0", features = ["attributes"] }
+anyhow = "1"
